add-viewlib: rename mur-setting-mgr to match 'settings' use in other modules
